requirements.txt
pandas
lxml
tensorflow
matplotlib
pillow
tensorflow-object-detection-api